move equals, reset_index
