+++ python/cudf/cudf/tests/test_series.py
@@ -10,7 +10,6 @@ import pyarrow as pa
 import pytest
 
 import cudf
-from cudf.api.extensions import no_default
 from cudf.core._compat import PANDAS_CURRENT_SUPPORTED_VERSION, PANDAS_VERSION
 from cudf.core.column.column import as_column
 from cudf.errors import MixedTypeError
@@ -71,164 +70,6 @@ def test_series_error_equality(sr1, sr2, op):
     gsr2 = cudf.from_pandas(sr2)
 
     assert_exceptions_equal(op, op, ([sr1, sr2],), ([gsr1, gsr2],))
-
-
-def test_set_index_unequal_length():
-    s = cudf.Series(dtype="float64")
-    with pytest.raises(ValueError):
-        s.index = [1, 2, 3]
-
-
-@pytest.mark.parametrize(
-    "lhs, rhs", [("a", "a"), ("a", "b"), (1, 1.0), (None, None), (None, "a")]
-)
-def test_equals_names(lhs, rhs):
-    lhs = cudf.Series([1, 2], name=lhs)
-    rhs = cudf.Series([1, 2], name=rhs)
-
-    got = lhs.equals(rhs)
-    expect = lhs.to_pandas().equals(rhs.to_pandas())
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data", [[True, False, None, True, False], [None, None], []]
-)
-@pytest.mark.parametrize("bool_dtype", ["bool", "boolean", pd.BooleanDtype()])
-def test_nullable_bool_dtype_series(data, bool_dtype):
-    psr = pd.Series(data, dtype=pd.BooleanDtype())
-    gsr = cudf.Series(data, dtype=bool_dtype)
-
-    assert_eq(psr, gsr.to_pandas(nullable=True))
-
-
-@pytest.mark.parametrize("level", [None, 0, "l0", 1, ["l0", 1]])
-@pytest.mark.parametrize("drop", [True, False])
-@pytest.mark.parametrize("original_name", [None, "original_ser"])
-@pytest.mark.parametrize("name", [None, "ser", no_default])
-@pytest.mark.parametrize("inplace", [True, False])
-def test_reset_index(level, drop, inplace, original_name, name):
-    midx = pd.MultiIndex.from_tuples(
-        [("a", 1), ("a", 2), ("b", 1), ("b", 2)], names=["l0", None]
-    )
-    ps = pd.Series(range(4), index=midx, name=original_name)
-    gs = cudf.from_pandas(ps)
-
-    if not drop and inplace:
-        pytest.skip(
-            "For exception checks, see "
-            "test_reset_index_dup_level_name_exceptions"
-        )
-
-    expect = ps.reset_index(level=level, drop=drop, name=name, inplace=inplace)
-
-    got = gs.reset_index(level=level, drop=drop, name=name, inplace=inplace)
-    if inplace:
-        expect = ps
-        got = gs
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize("level", [None, 0, 1, [None]])
-@pytest.mark.parametrize("drop", [False, True])
-@pytest.mark.parametrize("inplace", [False, True])
-@pytest.mark.parametrize("original_name", [None, "original_ser"])
-@pytest.mark.parametrize("name", [None, "ser"])
-def test_reset_index_dup_level_name(level, drop, inplace, original_name, name):
-    # midx levels are named [None, None]
-    midx = pd.MultiIndex.from_tuples([("a", 1), ("a", 2), ("b", 1), ("b", 2)])
-    ps = pd.Series(range(4), index=midx, name=original_name)
-    gs = cudf.from_pandas(ps)
-    if level == [None] or not drop and inplace:
-        pytest.skip(
-            "For exception checks, see "
-            "test_reset_index_dup_level_name_exceptions"
-        )
-
-    expect = ps.reset_index(level=level, drop=drop, inplace=inplace, name=name)
-    got = gs.reset_index(level=level, drop=drop, inplace=inplace, name=name)
-    if inplace:
-        expect = ps
-        got = gs
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize("drop", [True, False])
-@pytest.mark.parametrize("inplace", [True, False])
-@pytest.mark.parametrize("original_name", [None, "original_ser"])
-@pytest.mark.parametrize("name", [None, "ser"])
-def test_reset_index_named(drop, inplace, original_name, name):
-    ps = pd.Series(range(4), index=["x", "y", "z", "w"], name=original_name)
-    gs = cudf.from_pandas(ps)
-
-    ps.index.name = "cudf"
-    gs.index.name = "cudf"
-
-    if not drop and inplace:
-        pytest.skip(
-            "For exception checks, see "
-            "test_reset_index_dup_level_name_exceptions"
-        )
-
-    expect = ps.reset_index(drop=drop, inplace=inplace, name=name)
-    got = gs.reset_index(drop=drop, inplace=inplace, name=name)
-
-    if inplace:
-        expect = ps
-        got = gs
-
-    assert_eq(expect, got)
-
-
-def test_reset_index_dup_level_name_exceptions():
-    midx = pd.MultiIndex.from_tuples([("a", 1), ("a", 2), ("b", 1), ("b", 2)])
-    ps = pd.Series(range(4), index=midx)
-    gs = cudf.from_pandas(ps)
-
-    # Should specify duplicate level names with level number.
-    assert_exceptions_equal(
-        lfunc=ps.reset_index,
-        rfunc=gs.reset_index,
-        lfunc_args_and_kwargs=(
-            [],
-            {"level": [None]},
-        ),
-        rfunc_args_and_kwargs=(
-            [],
-            {"level": [None]},
-        ),
-    )
-
-    # Cannot use drop=False and inplace=True to turn a series into dataframe.
-    assert_exceptions_equal(
-        lfunc=ps.reset_index,
-        rfunc=gs.reset_index,
-        lfunc_args_and_kwargs=(
-            [],
-            {"drop": False, "inplace": True},
-        ),
-        rfunc_args_and_kwargs=(
-            [],
-            {"drop": False, "inplace": True},
-        ),
-    )
-
-    # Pandas raises the above exception should these two inputs crosses.
-    assert_exceptions_equal(
-        lfunc=ps.reset_index,
-        rfunc=gs.reset_index,
-        lfunc_args_and_kwargs=(
-            [],
-            {"level": [None], "drop": False, "inplace": True},
-        ),
-        rfunc_args_and_kwargs=(
-            [],
-            {"level": [None], "drop": False, "inplace": True},
-        ),
-    )
 
 
 def test_series_add_prefix():
